cositas

+++ CPP_07/ex01/hdr/Iter.hpp
@@ -6,7 +6,7 @@
 /*   By: ncastell <ncastell@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/23 19:59:10 by ncastell          #+#    #+#             */
-/*   Updated: 2024/10/28 15:55:42 by ncastell         ###   ########.fr       */
+/*   Updated: 2024/11/05 14:41:46 by ncastell         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@
 template <class T>
 void	iter(T *array, int len, void (*fun)(T&))
 {
-	for (int i = 0; i < len; ++i)
+	for (size_t i = 0; i < len; ++i)
 	fun(array[i]);
 }
